feat(snap): rename java pkg's
