+++ rskj-core/src/integrationTest/java/co/rsk/pte/PteIntegrationTest.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package pte;
+package co.rsk.pte;
 
 import co.rsk.util.OkHttpClientTestFixture;
 import co.rsk.util.cli.CommandLineFixture;
